+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-novn-buy-and-sell.spec.ts
@@ -145,19 +145,23 @@ describe('PortfolioCalculator', () => {
         valueWithCurrencyEffect: 0
       });
 
+      /**
+       * Closing price on 2022-03-07 is unknown,
+       * hence it uses the last unit price (2022-04-11): 87.8
+       */
       expect(portfolioSnapshot.historicalData[1]).toEqual({
         date: '2022-03-07',
         investmentValueWithCurrencyEffect: 151.6,
-        netPerformance: 0,
-        netPerformanceInPercentage: 0,
-        netPerformanceInPercentageWithCurrencyEffect: 0,
-        netPerformanceWithCurrencyEffect: 0,
-        netWorth: 151.6,
+        netPerformance: 24, // 2 * (87.8 - 75.8) = 24
+        netPerformanceInPercentage: 0.158311345646438, // 24 ÷ 151.6 = 0.158311345646438
+        netPerformanceInPercentageWithCurrencyEffect: 0.158311345646438, // 24 ÷ 151.6 = 0.158311345646438
+        netPerformanceWithCurrencyEffect: 24,
+        netWorth: 175.6, // 2 * 87.8 = 175.6
         totalAccountBalance: 0,
         totalInvestment: 151.6,
         totalInvestmentValueWithCurrencyEffect: 151.6,
-        value: 151.6,
-        valueWithCurrencyEffect: 151.6
+        value: 175.6, // 2 * 87.8 = 175.6
+        valueWithCurrencyEffect: 175.6
       });
 
       expect(