+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed an issue in the performance calculation on the date of an activity when the unit price differs from the market price
 - Fixed the horizontal overflow in the table of the benchmark component
 
 ## 2.160.0 - 2025-05-04